+++ public/locales/es/common.json
@@ -205,8 +205,10 @@
       "trade": "Intercambia"
     },
     "hero": {
-      "headline": "Genera Dai usando cripto como colateral",
-      "subheader": "Abre un Vault de Maker y genera Dai depositando tu colateral. Conecta tu cartera para comenzar"
+      "headlinePreAssetLandingPages": "Genera Dai usando cripto como colateral",
+      "subheaderPreAssetLandingPages": "Abre un Vault de Maker y genera Dai depositando tu colateral. Conecta tu cartera para comenzar",
+      "headline": "Borrow and multiply using your favourite cryptos",
+      "subheader": "ETH, WBTC, +30 more assets available to put to work from just 1.50% per year.<0/> Connect a wallet to get started"
     },
     "cards": {
       "dai": {
